fixing propTypes, refactoring history, adding ui
